requirements.txt
aiofiles==0.8.0
anyio==3.5.0
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
fastapi==0.71.0
h11==0.12.0
httptools==0.3.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.0
ipykernel==6.7.0
ipython==7.31.1
jedi==0.18.1
jupyter-client==7.1.1
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mpmath==1.2.1
Naked==0.1.31
nest-asyncio==1.5.4
numpy==1.21.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycryptodome==3.12.0
pydantic==1.9.0
pyDes==2.0.1
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
sympy==1.9
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
zipp==3.7.0
